+++ Runtime/ArConfig.cs
@@ -43,20 +43,6 @@ namespace UnityEngine.XR.ARCore
         public static ArConfig FromIntPtr(IntPtr value) => new ArConfig(value);
 
         /// <summary>
-        /// Represents a null <see cref="ArConfig"/>, i.e., one whose underlying native pointer is `null`.
-        /// This property is deprecated. Use <see langword="default"/> instead.
-        /// </summary>
-        [Obsolete("Use default instead.")]
-        public static ArConfig Null => default;
-
-        /// <summary>
-        /// (Read Only) Indicates whether this <see cref="ArConfig"/> is `null`.
-        /// This property is deprecated. Use the equality operator (`==`) to compare with `null` instead.
-        /// </summary>
-        [Obsolete("Compare to null instead.")]
-        public bool IsNull => m_Self == IntPtr.Zero;
-
-        /// <summary>
         /// Gets the underlying native pointer for this <see cref="ArConfig"/>.
         /// </summary>
         /// <returns>Returns the underlying native pointer for this <see cref="ArConfig"/>.</returns>